+++ src/euphorie/client/api/risk.py
@@ -1,3 +1,4 @@
+from Acquisition import aq_parent
 from zope.component import adapts
 from five import grok
 from euphorie.content.risk import evaluation_algorithm
@@ -31,31 +32,32 @@ class View(JsonView):
     check_update = True
 
     def do_GET(self):
-        self.risk = self.request.survey.restrictedTraverse(
+        self.risk = risk = self.request.survey.restrictedTraverse(
                 self.context.zodb_path.split('/'))
         info = {'id': self.context.id,
                 'type': 'risk',
-                'title': self.risk.title,
-                'problem-description': self.risk.problem_description,
-                'show-not-applicable': self.risk.show_notapplicable,
-                'evaluation-method': self.risk.evaluation_method,
+                'title': risk.title,
+                'module-title': aq_parent(risk).title,
+                'problem-description': risk.problem_description,
+                'show-not-applicable': risk.show_notapplicable,
+                'evaluation-method': risk.evaluation_method,
                 'present': self.context.identification,
                 'priority': self.context.priority,
                 'comment': self.context.comment,
                 }
         images = filter(None,
-                [export_image(self.risk, self.request,
+                [export_image(risk, self.request,
                     'image%s' % postfix, 'caption%s' % postfix,
                     width=150, height=500, direction='thumbnail')
                     for postfix in ['', '2', '3', '4']])
         if images:
             info['images'] = images
-        if HasText(self.risk.description):
-            info['description'] = self.risk.description
-        if HasText(self.risk.legal_reference):
-            info['legal-reference'] = self.risk.legal_reference
-        if self.risk.evaluation_method == 'calculated':
-            algorithm = evaluation_algorithm(self.risk)
+        if HasText(risk.description):
+            info['description'] = risk.description
+        if HasText(risk.legal_reference):
+            info['legal-reference'] = risk.legal_reference
+        if risk.evaluation_method == 'calculated':
+            algorithm = evaluation_algorithm(risk)
             info['evaluation-algorithm'] = algorithm
             if algorithm == 'french':
                 field = IFrenchEvaluation['default_severity']