+++ src/euphorie/client/api/tests/test_risk.py
@@ -33,6 +33,7 @@ class ViewTests(EuphorieFunctionalTestCase):
         from euphorie.client.model import Risk
         self.loginAsPortalOwner()
         (account, survey, survey_session) = _setup_session(self.portal)
+        survey['1'].title = u'Module title.'
         risk = survey['1']['2']
         risk.title = u'Everything is under control.'
         risk.problem_description = u'Not everything under control.'
@@ -46,14 +47,13 @@ class ViewTests(EuphorieFunctionalTestCase):
         response = view.do_GET()
         self.assertEqual(
                 set(response),
-                set(['id', 'type', 'title', 'problem-description',
-                     'show-not-applicable', 'evaluation-method',
-                     'present', 'priority', 'comment']))
+                set(['id', 'type', 'title', 'module-title',
+                      'problem-description', 'show-not-applicable',
+                      'evaluation-method', 'present', 'priority', 'comment']))
         self.assertEqual(response['id'], 2)
         self.assertEqual(response['type'], 'risk')
-        self.assertEqual(
-                response['title'],
-                u'Everything is under control.')
+        self.assertEqual(response['title'], u'Everything is under control.')
+        self.assertEqual(response['module-title'], u'Module title.')
         self.assertEqual(
                 response['problem-description'],
                 u'Not everything under control.')
@@ -84,11 +84,10 @@ class ViewTests(EuphorieFunctionalTestCase):
         response = view.do_GET()
         self.assertEqual(
                 set(response),
-                set(['id', 'type', 'title', 'problem-description',
-                     'show-not-applicable', 'evaluation-method',
-                     'present', 'priority', 'comment',
-                     'description', 'legal-reference',
-                     'evaluation-algorithm',
+                set(['id', 'type', 'title', 'module-title',
+                     'problem-description', 'show-not-applicable',
+                     'evaluation-method', 'present', 'priority', 'comment',
+                     'description', 'legal-reference', 'evaluation-algorithm',
                      'frequency', 'frequency-options',
                      'effect', 'effect-options',
                      'probability', 'probability-options',